Map Editor
Currently the editor is directly in the Game scene. In the future it would be better to move it to its own scene, with its own UI.

Controls:
Q - draw tiles
, and . - change active tile
- and = - change brush size

W - place armies
[ and ] - change active army

E - place city

Delete - clear tile contents

X, Y, P - toggle X, Y and Point symmetry

R - resize the TileMap (currently set to a hardcoded 200x200 size)

Most of these keyboard shortcuts are meant to be replaced by an actual UI in the future.

To change the number, names and gold of the players, and to change the names and descriptions of cities, you have to edit the .json file.