Color mask shader (URP)
Unit jsons now have an additional field: "maskTexture": "{path to the color mask texture}".

The color mask texture uses the R channel to indicate how much a given pixel in the unit's main texture is multiplied with the player's color. Channels G, B and A of the mask texture are currently unused.